rebases branch and makes small style enhancements
