+++ src/pages/NFTs/Collection/CollectionV2.tsx
@@ -355,7 +355,6 @@ const FiltersContainer: FC<{
 
   return (
     <NFT_FILTERS_CONTAINER index={displayIndex}>
-      {DisplayFilterIcon}
       <div className="flitersFlexContainer">
         {DisplayFilterIcon}
